Refactor `proto::message::AddIdentity` etc. to module
